move provider to separate class
return disposables

add missing git files
