+++ src/commands/configureLaunchDarkly.ts
@@ -10,18 +10,20 @@ export default async function configureLaunchDarkly(
 	config: Configuration,
 	api: LaunchDarklyAPI,
 	flagStore?: FlagStore,
+	disposables?: Array<Disposable>,
 ) {
 	const configureExtension: Disposable = commands.registerCommand('extension.configureLaunchDarkly', async () => {
 		try {
-			const configurationMenu = new ConfigurationMenu(config, api, ctx);
+			const configurationMenu = new ConfigurationMenu(config, api, ctx, disposables);
 			await configurationMenu.configure();
 			if (typeof flagStore === 'undefined') {
 				flagStore = new FlagStore(config, api);
 			} else {
 				await flagStore.reload();
 			}
-
-			await setupComponents(api, config, ctx, flagStore);
+			//console.log("here")
+			//console.log(disposables)
+			//await setupComponents(api, config, ctx, flagStore, disposables);
 			await ctx.globalState.update('LDConfigured', true);
 			window.showInformationMessage('LaunchDarkly configured successfully');
 		} catch (err) {
