+++ src/commands/createFlag.ts
@@ -3,10 +3,13 @@ import { LaunchDarklyAPI } from '../api';
 import { Configuration } from '../configuration';
 import { CreateFlagMenu } from '../createFlagMenu';
 
-export default async function createFlagCmd(ctx: ExtensionContext, config: Configuration, api: LaunchDarklyAPI) {
-	const createFlagCmd: Disposable = commands.registerCommand('launchdarkly.createFlag', async () => {
+export default function createFlagCmd(ctx: ExtensionContext, config: Configuration, api: LaunchDarklyAPI): Disposable {
+	const createFlagCmd = commands.registerCommand('launchdarkly.createFlag', async () => {
 		const configurationMenu = new CreateFlagMenu(config, api);
 		await configurationMenu.collectInputs();
 	});
+
 	ctx.subscriptions.push(createFlagCmd);
+
+	return createFlagCmd;
 }
