+++ src/providers.ts
@@ -2,40 +2,26 @@ import {
 	commands,
 	languages,
 	window,
-	CompletionItem,
-	CompletionItemKind,
-	CompletionItemProvider,
-	DocumentFilter,
 	ExtensionContext,
 	Position,
-	Range,
 	TextDocument,
 	workspace,
 	ConfigurationChangeEvent,
 	OutputChannel,
 } from 'vscode';
-import * as url from 'url';
-import opn = require('opn');
-import { kebabCase } from 'lodash';
+
 import { Configuration } from './configuration';
 import { LaunchDarklyAPI } from './api';
-import { FeatureFlagConfig } from './models';
 import { FlagStore } from './flagStore';
-import { FlagAliases } from './providers/codeRefs';
-import { LaunchDarklyHoverProvider } from './providers/hover';
 import { FlagNode } from './providers/flagListView';
 import { ClientSideEnable, refreshDiagnostics, subscribeToDocumentChanges } from './providers/diagnostics';
 import PubNub from 'pubnub';
 import globalClearCmd from './commands/clearGlobalContext';
 import configureLaunchDarkly from './commands/configureLaunchDarkly';
 import { extensionReload, setupComponents } from './utils';
-import createFlagCmd from './commands/createFlag';
+import generalCommands from './commands/generalCommands';
 
-const STRING_DELIMETERS = ['"', "'", '`'];
 export const FLAG_KEY_REGEX = /[A-Za-z0-9][.A-Za-z_\-0-9]*/;
-const LD_MODE: DocumentFilter = {
-	scheme: 'file',
-};
 
 type DebuggerWindows = Record<string, OutputChannel>;
 
@@ -47,75 +33,24 @@ export async function register(
 ): Promise<void> {
 	let aliases;
 
-	await configureLaunchDarkly(ctx, config, api, flagStore);
-	await createFlagCmd(ctx, config, api);
-
-	ctx.subscriptions.push(
-		commands.registerCommand('launchdarkly.toggleFlagContext', async () => {
-			try {
-				const key = ctx.workspaceState.get('LDFlagKey') as string;
-				if (key) {
-					const env = await flagStore.getFeatureFlag(key);
-					await api.patchFeatureFlagOn(config.project, key, !env.config.on);
-				}
-			} catch (err) {
-				window.showErrorMessage(`Could not patch flag: ${err.message}`);
-			}
-		}),
-		commands.registerTextEditorCommand('extension.openInLaunchDarkly', async (editor) => {
-			const flagKey = editor.document.getText(
-				editor.document.getWordRangeAtPosition(editor.selection.anchor, FLAG_KEY_REGEX),
-			);
-			if (!flagKey) {
-				window.showErrorMessage(
-					'[LaunchDarkly] Error retrieving flag (current cursor position is not a feature flag).',
-				);
-				return;
-			}
-
-			if (!config.accessToken) {
-				window.showErrorMessage('[LaunchDarkly] accessToken is not set.');
-				return;
-			}
-
-			if (!config.project) {
-				window.showErrorMessage('[LaunchDarkly] project is not set.');
-				return;
-			}
-
-			try {
-				const fKey = ctx.workspaceState.get('LDFlagKey') as string;
-				await openFlagInBrowser(config, fKey, flagStore);
-			} catch (err) {
-				let errMsg = `Encountered an unexpected error retrieving the flag ${flagKey}`;
-				if (err.statusCode == 404) {
-					// Try resolving the flag key to kebab case
-					try {
-						await openFlagInBrowser(config, kebabCase(flagKey), flagStore);
-						return;
-					} catch (err) {
-						if (err.statusCode == 404) {
-							errMsg = `Could not find the flag ${flagKey}`;
-						}
-					}
-				}
-				console.error(`Failed opening browser: ${err}`);
-				window.showErrorMessage(`[LaunchDarkly] ${errMsg}`);
-			}
-		}),
-	);
 	await globalClearCmd(ctx, config);
+
+	let disposables = await generalCommands(ctx, config, api, flagStore);
+	console.log('about to configure');
+	console.log(disposables);
+	await configureLaunchDarkly(ctx, config, api, flagStore, disposables);
+
 	// Handle manual changes to extension configuration
 	workspace.onDidChangeConfiguration(async (e: ConfigurationChangeEvent) => {
 		if (e.affectsConfiguration('launchdarkly')) {
-			extensionReload(config, ctx);
+			disposables = await extensionReload(config, ctx, disposables);
 		}
 	});
 	if (!config.isConfigured) {
 		return;
 	}
 	if (typeof flagStore !== 'undefined') {
-		await setupComponents(api, config, ctx, flagStore, aliases);
+		await setupComponents(api, config, ctx, flagStore);
 
 		const Lddebugger = window.createOutputChannel('LaunchDarkly All');
 		const LddIdentify = window.createOutputChannel('LaunchDarkly Identify Debug');
@@ -152,13 +87,6 @@ export async function register(
 			commands.registerCommand('launchdarkly.OpenFlag', (node: FlagNode) =>
 				window.activeTextEditor.revealRange(node.range),
 			),
-			languages.registerCompletionItemProvider(
-				LD_MODE,
-				new LaunchDarklyCompletionItemProvider(config, flagStore, aliases),
-				"'",
-				'"',
-			),
-			languages.registerHoverProvider(LD_MODE, new LaunchDarklyHoverProvider(config, flagStore, ctx, aliases)),
 		);
 	}
 
@@ -374,72 +302,7 @@ const parseFeature = (event: FeatureEvent) => {
 
 	return line;
 };
-class LaunchDarklyCompletionItemProvider implements CompletionItemProvider {
-	private readonly flagStore: FlagStore;
-	private readonly config: Configuration;
-	private readonly aliases?: FlagAliases;
 
-	constructor(config: Configuration, flagStore: FlagStore, aliases?: FlagAliases) {
-		this.config = config;
-		this.flagStore = flagStore;
-		this.aliases = aliases;
-	}
-
-	public provideCompletionItems(document: TextDocument, position: Position): Thenable<CompletionItem[]> {
-		if (isPrecedingCharStringDelimiter(document, position)) {
-			// eslint-disable-next-line no-async-promise-executor
-			return new Promise(async (resolve) => {
-				if (this.config.enableAutocomplete) {
-					const flags = await this.flagStore.allFlagsMetadata();
-					const flagCompletes = [];
-					// eslint-disable-next-line @typescript-eslint/no-unused-vars
-					for (const [key, flag] of Object.entries(flags)) {
-						const flagCompletion = new CompletionItem(flag.key, CompletionItemKind.Field);
-						flagCompletion.detail = flag.description ? flag.description : '';
-						flagCompletes.push(flagCompletion);
-					}
-					resolve(flagCompletes);
-					return;
-				}
-				resolve(null);
-			});
-		}
-	}
+export function isPrecedingCharStringDelimiter(document: TextDocument, pos: Position): any {
+	throw new Error('Function not implemented.');
 }
-
-const openFlagInBrowser = async (config: Configuration, flagKey: string, flagStore: FlagStore) => {
-	const { flag } = await flagStore.getFeatureFlag(flagKey);
-
-	// Default to first environment
-	let env: FeatureFlagConfig = Object.values(flag.environments)[0];
-	let sitePath = env._site.href;
-
-	if (!config.env) {
-		window.showWarningMessage('[LaunchDarkly] env is not set. Falling back to first environment.');
-	} else if (!flag.environments[config.env]) {
-		window.showWarningMessage(
-			`[LaunchDarkly] Configured environment '${config.env}' has been deleted. Falling back to first environment.`,
-		);
-	} else {
-		env = flag.environments[config.env];
-		sitePath = env._site.href;
-	}
-	opn(url.resolve(config.baseUri, sitePath));
-};
-
-export function isPrecedingCharStringDelimiter(document: TextDocument, position: Position): boolean {
-	const range = document.getWordRangeAtPosition(position, FLAG_KEY_REGEX);
-	if (!range || !range.start || range.start.character === 0) {
-		return false;
-	}
-	const c = new Range(
-		range.start.line,
-		candidateTextStartLocation(range.start.character),
-		range.start.line,
-		range.start.character,
-	);
-	const candidate = document.getText(c).trim().replace('(', '');
-	return STRING_DELIMETERS.indexOf(candidate) !== -1;
-}
-
-const candidateTextStartLocation = (char: number) => (char === 1 ? 0 : char - 2);
