+++ src/utils.ts
@@ -1,11 +1,15 @@
-import { commands, ExtensionContext, languages, window } from 'vscode';
+import { time } from 'console';
+import { commands, Disposable, DocumentFilter, ExtensionContext, languages, window } from 'vscode';
 import { LaunchDarklyAPI } from './api';
+import generalCommands from './commands/generalCommands';
 import { Configuration } from './configuration';
 import { FlagStore } from './flagStore';
 import { FlagAliases } from './providers/codeRefs';
+import LaunchDarklyCompletionItemProvider from './providers/completion';
 import { FlagCodeLensProvider } from './providers/flagLens';
 import { LaunchDarklyFlagListProvider } from './providers/flagListView';
 import { LaunchDarklyTreeViewProvider } from './providers/flagsView';
+import { LaunchDarklyHoverProvider } from './providers/hover';
 import { LaunchDarklyMetricsTreeViewProvider } from './providers/metricsView';
 import { QuickLinksListProvider } from './providers/quickLinksView';
 
@@ -17,12 +21,17 @@ export default async function checkExistingCommand(commandName: string): Promise
 	return false;
 }
 
-export async function extensionReload(config: Configuration, ctx: ExtensionContext) {
+export async function extensionReload(
+	config: Configuration,
+	ctx: ExtensionContext,
+	disposables?: Array<Disposable>,
+): Promise<Array<Disposable>> {
 	// Read in latest version of config
 	await config.reload();
 	const newApi = new LaunchDarklyAPI(config);
 	const flagStore = new FlagStore(config, newApi);
-	await setupComponents(newApi, config, ctx, flagStore);
+	await setupComponents(newApi, config, ctx, flagStore, disposables);
+	return generalCommands(ctx, config, newApi, flagStore);
 }
 
 export async function setupComponents(
@@ -30,8 +39,19 @@ export async function setupComponents(
 	config: Configuration,
 	ctx: ExtensionContext,
 	flagStore: FlagStore,
-	aliases?: FlagAliases,
+	disposables?: Array<Disposable>,
 ) {
+	console.log(`setting up ${Date.now()}`);
+	console.log(disposables);
+	if (disposables) {
+		disposables.map((item) => {
+			item.dispose();
+		});
+	}
+	const waitTill = new Date(new Date().getTime() + 5 * 1000);
+	// eslint-disable-next-line no-empty
+	while (waitTill > new Date()) {}
+
 	// Add metrics view
 	const metricsView = new LaunchDarklyMetricsTreeViewProvider(api, config, ctx);
 	window.registerTreeDataProvider('launchdarklyMetrics', metricsView);
@@ -40,6 +60,7 @@ export async function setupComponents(
 	const quickLinksView = new QuickLinksListProvider(config, flagStore);
 	window.registerTreeDataProvider('launchdarklyQuickLinks', quickLinksView);
 
+	let aliases;
 	if (config.enableAliases) {
 		aliases = new FlagAliases(config, ctx);
 		if (aliases.codeRefsVersionCheck()) {
@@ -57,9 +78,20 @@ export async function setupComponents(
 	const codeLens = new FlagCodeLensProvider(api, config, flagStore, aliases);
 	const listView = new LaunchDarklyFlagListProvider(config, codeLens);
 	window.registerTreeDataProvider('launchdarklyFlagList', listView);
+
+	const LD_MODE: DocumentFilter = {
+		scheme: 'file',
+	};
 	ctx.subscriptions.push(
 		window.onDidChangeActiveTextEditor(listView.setFlagsinDocument),
 		languages.registerCodeLensProvider('*', codeLens),
+		languages.registerCompletionItemProvider(
+			LD_MODE,
+			new LaunchDarklyCompletionItemProvider(config, flagStore, aliases),
+			"'",
+			'"',
+		),
+		languages.registerHoverProvider(LD_MODE, new LaunchDarklyHoverProvider(config, flagStore, ctx, aliases)),
 	);
 
 	codeLens.start();
